+++ app/src/main/java/com/viettel/vpmt/mobiletv/network/dto/PartOfFilm.java
@@ -4,14 +4,14 @@ package com.viettel.vpmt.mobiletv.network.dto;
  * Created by ThanhTD on 3/28/2016.
  */
 public class PartOfFilm {
-    private float id;
+    private Integer id;
     private String name;
 
-    public float getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(float id) {
+    public void setId(Integer id) {
         this.id = id;
     }
